feat: Add shorten_string filter to custom_filters.py
The `custom_filters.py` file has been updated to include a new filter called `shorten_string`. This filter takes a string value and an argument, and returns the last specified number of characters of the value, preceded by an ellipsis. This addition enhances the functionality of the custom filters module and allows for more flexible string manipulation in templates.